Refactor Outlook search query parsing
