+++ apps/web/utils/outlook/message.ts
@@ -11,6 +11,11 @@ import { formatEmailWithName } from "@/utils/email";
 import type { Logger } from "@/utils/logger";
 import { isOutlookThrottlingError } from "@/utils/error";
 import { resolveMicrosoftGraphNextLink } from "@/utils/outlook/page-token";
+import {
+  matchesOutlookMetadataFilters,
+  parseOutlookMetadataSearchQuery,
+  stripStandaloneOutlookStateTerms,
+} from "@/utils/outlook/search-query";
 
 // Standard fields to select when fetching messages from Microsoft Graph API
 // internetMessageId is the RFC 5322 Message-ID header, needed for cross-provider email threading
@@ -172,12 +177,6 @@ const KQL_BOOLEAN_OPERATOR_PATTERN = /\b(?:AND|OR|NOT)\b/i;
 const KQL_BOOLEAN_PATTERN = /\b(?:AND|OR|NOT)\b/gi;
 const KQL_GROUPING_DETECTION_PATTERN = /[()]/;
 const KQL_GROUPING_PATTERN = /[()]/g;
-const OUTLOOK_CATEGORY_SEARCH_FIELDS = new Set([
-  "category",
-  "categories",
-  "label",
-  "labels",
-]);
 // Stripped only when degrading a failed KQL query to free-text fallback, since
 // comparison filters and read-state terms hurt keyword recall there.
 const OUTLOOK_COMPARISON_FILTER_PATTERN =
@@ -345,21 +344,6 @@ function collapseOutlookSearchQueryToText(query: string): string {
     .trim();
 }
 
-export function stripStandaloneOutlookStateTerms(query: string) {
-  return splitOutlookQueryTerms(query)
-    .filter((term) => {
-      const normalized = term.replace(/^[()]+|[()]+$/g, "").toLowerCase();
-      return normalized !== "read" && normalized !== "unread";
-    })
-    .join(" ");
-}
-
-export function getStandaloneOutlookStateTerms(query: string) {
-  return splitOutlookQueryTerms(query)
-    .map((term) => term.replace(/^[()]+|[()]+$/g, "").toLowerCase())
-    .filter((term) => term === "read" || term === "unread");
-}
-
 export function getOutlookComparisonFilters(query: string) {
   return Array.from(
     query.matchAll(OUTLOOK_COMPARISON_FILTER_PATTERN),
@@ -369,192 +353,6 @@ export function getOutlookComparisonFilters(query: string) {
 
 export function stripOutlookComparisonFilters(query: string) {
   return query.replace(OUTLOOK_COMPARISON_FILTER_PATTERN, " ");
-}
-
-function splitOutlookQueryTerms(query: string) {
-  const terms: string[] = [];
-  let current = "";
-  let inQuotes = false;
-
-  for (const char of query) {
-    if (char === '"') {
-      inQuotes = !inQuotes;
-      current += char;
-      continue;
-    }
-
-    if (!inQuotes && /\s/.test(char)) {
-      if (current) {
-        terms.push(current);
-        current = "";
-      }
-      continue;
-    }
-
-    current += char;
-  }
-
-  if (current) {
-    terms.push(current);
-  }
-
-  return terms;
-}
-
-type OutlookMetadataFilters = {
-  isRead?: boolean;
-  categoryNames: string[];
-};
-
-const EMPTY_METADATA_PARSE_RESULT = {
-  searchQuery: "",
-  filters: { isRead: undefined, categoryNames: [] as string[] },
-  odataFilters: [] as string[],
-};
-
-function parseOutlookMetadataSearchQuery(
-  query: string,
-  categoryMap: Map<string, string>,
-): {
-  searchQuery: string;
-  filters: OutlookMetadataFilters;
-  odataFilters: string[];
-} {
-  if (!query) return EMPTY_METADATA_PARSE_RESULT;
-
-  const stateTerms = getStandaloneOutlookStateTerms(query);
-  const hasRead = stateTerms.includes("read");
-  const hasUnread = stateTerms.includes("unread");
-  const isRead = hasRead === hasUnread ? undefined : hasRead;
-  const categoryLookup = createOutlookCategoryLookup(categoryMap);
-
-  const terms = splitOutlookQueryTerms(stripStandaloneOutlookStateTerms(query));
-  const remainingTerms: string[] = [];
-  const categoryNames: string[] = [];
-
-  for (const term of terms) {
-    const parsedField = parseOutlookFieldTerm(term);
-    if (parsedField && OUTLOOK_CATEGORY_SEARCH_FIELDS.has(parsedField.field)) {
-      const categoryName = categoryLookup.get(
-        normalizeOutlookMetadataSearchValue(parsedField.value),
-      );
-      if (categoryName) {
-        categoryNames.push(categoryName);
-        continue;
-      }
-    }
-
-    remainingTerms.push(term);
-  }
-
-  let searchQuery = remainingTerms.join(" ").trim();
-  if (searchQuery) {
-    const categoryName = categoryLookup.get(
-      normalizeOutlookMetadataSearchValue(unquoteSearchValue(searchQuery)),
-    );
-    if (categoryName) {
-      categoryNames.push(categoryName);
-      searchQuery = "";
-    }
-  }
-
-  const filters = {
-    isRead,
-    categoryNames: [...new Set(categoryNames)],
-  };
-
-  return {
-    searchQuery,
-    filters,
-    odataFilters: createOutlookMetadataODataFilters(filters),
-  };
-}
-
-function parseOutlookFieldTerm(term: string) {
-  const colonIndex = term.indexOf(":");
-  if (colonIndex === -1) return null;
-
-  const field = stripOutlookGrouping(term.slice(0, colonIndex))
-    .trim()
-    .toLowerCase();
-  const value = stripOutlookGrouping(term.slice(colonIndex + 1)).trim();
-  if (!field || !value || URL_SCHEME_PATTERN.test(term)) return null;
-
-  return {
-    field,
-    value: unquoteSearchValue(value),
-  };
-}
-
-function stripOutlookGrouping(value: string) {
-  return value.replace(/^[()]+|[()]+$/g, "");
-}
-
-function unquoteSearchValue(value: string) {
-  const trimmed = value.trim();
-  if (trimmed.length >= 2 && trimmed.startsWith('"') && trimmed.endsWith('"')) {
-    return trimmed.slice(1, -1);
-  }
-  return trimmed;
-}
-
-function createOutlookCategoryLookup(categoryMap: Map<string, string>) {
-  const lookup = new Map<string, string>();
-
-  for (const categoryName of categoryMap.keys()) {
-    const normalized = normalizeOutlookMetadataSearchValue(categoryName);
-    if (!normalized) continue;
-
-    lookup.set(normalized, categoryName);
-
-    const pluralAlias = `${normalized}s`;
-    if (!normalized.endsWith("s") && !lookup.has(pluralAlias)) {
-      lookup.set(pluralAlias, categoryName);
-    }
-  }
-
-  return lookup;
-}
-
-function normalizeOutlookMetadataSearchValue(value: string) {
-  return value.trim().replace(/\s+/g, " ").toLowerCase();
-}
-
-function createOutlookMetadataODataFilters(filters: OutlookMetadataFilters) {
-  const odataFilters: string[] = [];
-
-  if (typeof filters.isRead === "boolean") {
-    odataFilters.push(`isRead eq ${filters.isRead}`);
-  }
-
-  for (const categoryName of filters.categoryNames) {
-    odataFilters.push(
-      `categories/any(category: category eq '${escapeODataString(categoryName)}')`,
-    );
-  }
-
-  return odataFilters;
-}
-
-function matchesOutlookMetadataFilters(
-  message: Message,
-  filters: OutlookMetadataFilters,
-) {
-  if (
-    typeof filters.isRead === "boolean" &&
-    message.isRead !== filters.isRead
-  ) {
-    return false;
-  }
-
-  if (filters.categoryNames.length) {
-    const messageCategories = new Set(message.categories ?? []);
-    return filters.categoryNames.every((categoryName) =>
-      messageCategories.has(categoryName),
-    );
-  }
-
-  return true;
 }
 
 export async function queryBatchMessages(
